Fix (Step2): Winning init 코드 Generator 로 분리
